+++ Memoria.Launcher/Languages/Lang.cs
@@ -1,5 +1,4 @@
-﻿using Ini;
-using System;
+﻿using System;
 using System.Globalization;
 using System.Reflection;
 using System.Xml;
