+++ Memoria.Launcher/Launcher/Settings.cs
@@ -1,17 +1,7 @@
 using System;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Media;
-using Ini;
 using Application = System.Windows.Application;
-using Binding = System.Windows.Data.Binding;
-using ComboBox = System.Windows.Controls.ComboBox;
-using Control = System.Windows.Controls.Control;
-using HorizontalAlignment = System.Windows.HorizontalAlignment;
 
 // ReSharper disable UnusedMember.Local
 // ReSharper disable InconsistentNaming
