+++ Memoria.Launcher/Launcher/SettingsGrid_Vanilla.cs
@@ -1,4 +1,3 @@
-using Ini;
 using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
